+++ src/scip/cons_expr_exp.c
@@ -366,7 +366,7 @@ SCIP_DECL_CONSEXPR_EXPRHASH(hashExp)
    assert(expr2key != NULL);
    assert(hashkey != NULL);
 
-   *hashkey = (unsigned int)EXP_HASHKEY;
+   *hashkey = EXP_HASHKEY;
 
    assert(SCIPhashmapExists(expr2key, (void*)SCIPgetConsExprExprChildren(expr)[0]));
    childhash = (unsigned int)(size_t)SCIPhashmapGetImage(expr2key, SCIPgetConsExprExprChildren(expr)[0]);
